server/requirements.txt
aiohttp==3.8.1
fastapi==0.70.1
pyhumps==3.0.2
pydantic>=1.8.2
redis==4.0.2
uvicorn==0.16.0
